refactor(vsock-guest): split connection test support
