+++ crates/vsock-guest/tests/connection/support/mod.rs
@@ -1,0 +1,21 @@
+mod connection;
+mod exec;
+mod process;
+mod protocol;
+mod temp_paths;
+
+pub(crate) use connection::{finish_guest_connection, start_guest_connection};
+pub(crate) use exec::{
+    DRAIN_DEADLINE_SECS, LARGE_ENV_COMMAND, LONG_RUNNING_EXEC_TIMEOUT_MS, assert_large_env_stdout,
+    large_env_entries, large_env_values, read_exec_output_chunk, read_exec_result,
+    send_exec_cancel, send_exec_start, send_exec_start_request, send_exec_start_with_env,
+    stderr_data, stdout_data,
+};
+pub(crate) use process::{
+    OrphanProcessGuard, ProcessGroupFileGuard, orphan_sleep_command, pid_alive, wait_for_pid_exit,
+};
+pub(crate) use protocol::{
+    assert_ping_pong, read_and_discard_message, read_error_response, read_message,
+    send_control_payload, send_quiesce_operations, send_resume_operations,
+};
+pub(crate) use temp_paths::{unique_pid_path, unique_socket_path, unique_tmp_path};
