+++ crates/vsock-guest/tests/connection/support/process.rs
@@ -2,84 +2,9 @@ use std::ffi::CString;
 use std::os::fd::{AsRawFd, FromRawFd, OwnedFd};
 use std::os::unix::ffi::OsStrExt;
 use std::path::Path;
-use std::thread;
 use std::time::{Duration, Instant};
 
-use vsock_guest::handle_connection;
-use vsock_proto::{
-    self, ExecCapturedOutput, ExecOutputPolicy, ExecOutputStream, ExecTermination, MSG_ERROR,
-    MSG_EXEC_CANCEL, MSG_EXEC_OUTPUT, MSG_EXEC_RESULT, MSG_EXEC_START, MSG_PING, MSG_PONG,
-    MSG_QUIESCE_OPERATIONS, MSG_RESUME_OPERATIONS,
-};
-
-pub(crate) const DRAIN_DEADLINE_SECS: u64 = 5;
-pub(crate) const LONG_RUNNING_EXEC_TIMEOUT_MS: u32 = 60_000;
-pub(crate) const LARGE_ENV_COMMAND: &str =
-    "printf '%s:%s:%s:%s:%s\\n' \"$SMALL\" \"${#BIG_A}\" \"${#BIG_B}\" \"${#BIG_C}\" \"${#BIG_D}\"";
-
-pub(crate) struct TempPathGuard(String);
-
-impl TempPathGuard {
-    fn new(path: String) -> Self {
-        let _ = std::fs::remove_file(&path);
-        Self(path)
-    }
-
-    pub(crate) fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl Drop for TempPathGuard {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_file(&self.0);
-    }
-}
-
-pub(crate) fn unique_tmp_path(label: &str, suffix: &str) -> TempPathGuard {
-    let nonce = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_nanos();
-    TempPathGuard::new(format!(
-        "/tmp/vsock-test-{label}-{}-{nonce}{suffix}",
-        std::process::id(),
-    ))
-}
-
-pub(crate) fn unique_socket_path(label: &str) -> TempPathGuard {
-    unique_tmp_path(label, ".sock")
-}
-
-pub(crate) fn unique_pid_path(label: &str) -> TempPathGuard {
-    unique_tmp_path(label, ".pid")
-}
-
-pub(crate) fn large_env_values() -> [String; 4] {
-    [
-        "A".repeat(40 * 1024),
-        "B".repeat(40 * 1024),
-        "C".repeat(40 * 1024),
-        "D".repeat(40 * 1024),
-    ]
-}
-
-pub(crate) fn large_env_entries(values: &[String; 4]) -> [(&'static str, &str); 5] {
-    [
-        ("SMALL", "ok"),
-        ("BIG_A", values[0].as_str()),
-        ("BIG_B", values[1].as_str()),
-        ("BIG_C", values[2].as_str()),
-        ("BIG_D", values[3].as_str()),
-    ]
-}
-
-pub(crate) fn assert_large_env_stdout(stdout: &[u8]) {
-    assert_eq!(
-        String::from_utf8_lossy(stdout),
-        "ok:40960:40960:40960:40960\n"
-    );
-}
+use super::temp_paths::{TempPathGuard, unique_pid_path};
 
 pub(crate) struct OrphanProcessGuard {
     pid_file: TempPathGuard,
@@ -117,271 +42,6 @@ impl Drop for OrphanProcessGuard {
 
 pub(crate) fn orphan_sleep_command(marker: &str, pid_path: &str) -> String {
     format!("sleep 30 & echo $! > {pid_path}; echo {marker}")
-}
-
-pub(crate) fn read_message(stream: &mut impl std::io::Read) -> vsock_proto::RawMessage {
-    let mut hdr = [0u8; 4];
-    stream.read_exact(&mut hdr).unwrap();
-    let body_len = u32::from_be_bytes(hdr) as usize;
-    let mut body = vec![0u8; body_len];
-    stream.read_exact(&mut body).unwrap();
-
-    let mut full = Vec::with_capacity(4 + body_len);
-    full.extend_from_slice(&hdr);
-    full.extend_from_slice(&body);
-    let mut decoder = vsock_proto::Decoder::new();
-    let msgs = decoder.decode(&full).unwrap();
-    assert_eq!(msgs.len(), 1);
-    msgs.into_iter().next().unwrap()
-}
-
-/// Read one framed message from the stream and discard it.
-pub(crate) fn read_and_discard_message(stream: &mut impl std::io::Read) {
-    let _ = read_message(stream);
-}
-
-#[derive(Debug)]
-pub(crate) struct ExecOutputChunk {
-    pub(crate) stream: ExecOutputStream,
-    pub(crate) output_seq: u32,
-    pub(crate) chunk: Vec<u8>,
-    pub(crate) truncated: bool,
-}
-
-#[derive(Debug)]
-pub(crate) struct ExecResult {
-    pub(crate) termination: ExecTermination,
-    pub(crate) stdout: Option<Vec<u8>>,
-    pub(crate) stderr: Option<Vec<u8>>,
-    pub(crate) stdout_truncated: bool,
-    pub(crate) stderr_truncated: bool,
-    pub(crate) diagnostic: String,
-}
-
-pub(crate) fn start_guest_connection() -> (thread::JoinHandle<()>, std::os::unix::net::UnixStream) {
-    let (guest_stream, mut host_stream) = std::os::unix::net::UnixStream::pair().unwrap();
-    let handle = thread::spawn(move || {
-        let _ = handle_connection(guest_stream);
-    });
-    read_and_discard_message(&mut host_stream);
-    host_stream
-        .set_read_timeout(Some(Duration::from_secs(5)))
-        .unwrap();
-    (handle, host_stream)
-}
-
-pub(crate) fn finish_guest_connection(
-    handle: thread::JoinHandle<()>,
-    host_stream: std::os::unix::net::UnixStream,
-) {
-    drop(host_stream);
-    let _ = handle.join();
-}
-
-pub(crate) fn send_exec_start(
-    stream: &mut impl std::io::Write,
-    seq: u32,
-    command: &str,
-    timeout_ms: u32,
-    stdout: ExecOutputPolicy,
-    stderr: ExecOutputPolicy,
-) {
-    send_exec_start_with_env(stream, seq, command, timeout_ms, &[], stdout, stderr);
-}
-
-pub(crate) fn send_exec_start_with_env(
-    stream: &mut impl std::io::Write,
-    seq: u32,
-    command: &str,
-    timeout_ms: u32,
-    env: &[(&str, &str)],
-    stdout: ExecOutputPolicy,
-    stderr: ExecOutputPolicy,
-) {
-    let payload =
-        vsock_proto::encode_exec_start(timeout_ms, command, env, false, "test", stdout, stderr)
-            .unwrap();
-    let msg = vsock_proto::encode(MSG_EXEC_START, seq, &payload).unwrap();
-    stream.write_all(&msg).unwrap();
-}
-
-pub(crate) fn send_exec_start_request(
-    stream: &mut impl std::io::Write,
-    seq: u32,
-    request: vsock_proto::ExecStartEncodeRequest<'_>,
-) {
-    let payload = vsock_proto::encode_exec_start_with_expected_exit_codes(request).unwrap();
-    let msg = vsock_proto::encode(MSG_EXEC_START, seq, &payload).unwrap();
-    stream.write_all(&msg).unwrap();
-}
-
-pub(crate) fn read_error_response(stream: &mut impl std::io::Read, seq: u32) -> String {
-    let msg = read_message(stream);
-    assert_eq!(msg.msg_type, MSG_ERROR);
-    assert_eq!(msg.seq, seq);
-    vsock_proto::decode_error(&msg.payload).unwrap().to_owned()
-}
-
-pub(crate) fn assert_ping_pong<T>(stream: &mut T, seq: u32)
-where
-    T: std::io::Read + std::io::Write,
-{
-    let ping = vsock_proto::encode(MSG_PING, seq, &[]).unwrap();
-    stream.write_all(&ping).unwrap();
-    let pong = read_message(stream);
-    assert_eq!(pong.msg_type, MSG_PONG);
-    assert_eq!(pong.seq, seq);
-    assert!(pong.payload.is_empty());
-}
-
-pub(crate) fn send_exec_cancel(stream: &mut impl std::io::Write, seq: u32) {
-    let payload = vsock_proto::encode_exec_cancel();
-    let msg = vsock_proto::encode(MSG_EXEC_CANCEL, seq, &payload).unwrap();
-    stream.write_all(&msg).unwrap();
-}
-
-pub(crate) fn send_empty_control(stream: &mut impl std::io::Write, msg_type: u8, seq: u32) {
-    let msg = vsock_proto::encode(msg_type, seq, &[]).unwrap();
-    stream.write_all(&msg).unwrap();
-}
-
-pub(crate) fn send_control_payload(
-    stream: &mut impl std::io::Write,
-    msg_type: u8,
-    seq: u32,
-    payload: &[u8],
-) {
-    let msg = vsock_proto::encode(msg_type, seq, payload).unwrap();
-    stream.write_all(&msg).unwrap();
-}
-
-pub(crate) fn send_quiesce_operations(stream: &mut impl std::io::Write, seq: u32) {
-    send_empty_control(stream, MSG_QUIESCE_OPERATIONS, seq);
-}
-
-pub(crate) fn send_resume_operations(stream: &mut impl std::io::Write, seq: u32) {
-    send_empty_control(stream, MSG_RESUME_OPERATIONS, seq);
-}
-
-pub(crate) fn read_exec_result(
-    stream: &mut impl std::io::Read,
-    seq: u32,
-) -> (Vec<ExecOutputChunk>, ExecResult) {
-    let mut decoder = vsock_proto::Decoder::new();
-    let mut buf = [0u8; 4096];
-    let mut chunks = Vec::new();
-    loop {
-        let n = read_retry_eintr(stream, &mut buf).unwrap();
-        assert!(n > 0, "unexpected EOF waiting for exec result");
-        for msg in decoder.decode(buf.get(..n).unwrap_or_default()).unwrap() {
-            if msg.seq != seq {
-                continue;
-            }
-            match msg.msg_type {
-                MSG_EXEC_OUTPUT => {
-                    let decoded = vsock_proto::decode_exec_output(&msg.payload).unwrap();
-                    chunks.push(ExecOutputChunk {
-                        stream: decoded.stream,
-                        output_seq: decoded.output_seq,
-                        chunk: decoded.chunk.to_vec(),
-                        truncated: decoded.truncated,
-                    });
-                }
-                MSG_EXEC_RESULT => {
-                    let decoded = vsock_proto::decode_exec_result(&msg.payload).unwrap();
-                    return (
-                        chunks,
-                        ExecResult {
-                            termination: decoded.termination,
-                            stdout: captured_to_vec(decoded.stdout),
-                            stderr: captured_to_vec(decoded.stderr),
-                            stdout_truncated: captured_truncated(decoded.stdout),
-                            stderr_truncated: captured_truncated(decoded.stderr),
-                            diagnostic: decoded.diagnostic.to_string(),
-                        },
-                    );
-                }
-                MSG_ERROR => {
-                    let error = vsock_proto::decode_error(&msg.payload).unwrap();
-                    panic!("unexpected exec operation error for seq={seq}: {error}");
-                }
-                other => panic!("unexpected exec operation response type: 0x{other:02X}"),
-            }
-        }
-    }
-}
-
-pub(crate) fn read_exec_output_chunk(stream: &mut impl std::io::Read, seq: u32) -> ExecOutputChunk {
-    let mut decoder = vsock_proto::Decoder::new();
-    let mut buf = [0u8; 4096];
-    loop {
-        let n = read_retry_eintr(stream, &mut buf).unwrap();
-        assert!(n > 0, "unexpected EOF waiting for exec output");
-        for msg in decoder.decode(buf.get(..n).unwrap_or_default()).unwrap() {
-            if msg.seq != seq {
-                continue;
-            }
-            match msg.msg_type {
-                MSG_EXEC_OUTPUT => {
-                    let decoded = vsock_proto::decode_exec_output(&msg.payload).unwrap();
-                    return ExecOutputChunk {
-                        stream: decoded.stream,
-                        output_seq: decoded.output_seq,
-                        chunk: decoded.chunk.to_vec(),
-                        truncated: decoded.truncated,
-                    };
-                }
-                MSG_EXEC_RESULT => panic!("unexpected exec result before output"),
-                MSG_ERROR => {
-                    let error = vsock_proto::decode_error(&msg.payload).unwrap();
-                    panic!("unexpected exec operation error for seq={seq}: {error}");
-                }
-                other => panic!("unexpected exec operation response type: 0x{other:02X}"),
-            }
-        }
-    }
-}
-
-fn captured_to_vec(captured: ExecCapturedOutput<'_>) -> Option<Vec<u8>> {
-    match captured {
-        ExecCapturedOutput::Discarded => None,
-        ExecCapturedOutput::Captured { bytes, .. } => Some(bytes.to_vec()),
-    }
-}
-
-fn captured_truncated(captured: ExecCapturedOutput<'_>) -> bool {
-    match captured {
-        ExecCapturedOutput::Discarded => false,
-        ExecCapturedOutput::Captured { truncated, .. } => truncated,
-    }
-}
-
-fn output_data(chunks: &[ExecOutputChunk], stream: ExecOutputStream) -> Vec<u8> {
-    chunks
-        .iter()
-        .filter(|chunk| chunk.stream == stream && !chunk.truncated)
-        .flat_map(|chunk| chunk.chunk.iter().copied())
-        .collect()
-}
-
-pub(crate) fn stdout_data(chunks: &[ExecOutputChunk]) -> Vec<u8> {
-    output_data(chunks, ExecOutputStream::Stdout)
-}
-
-pub(crate) fn stderr_data(chunks: &[ExecOutputChunk]) -> Vec<u8> {
-    output_data(chunks, ExecOutputStream::Stderr)
-}
-
-pub(crate) fn read_retry_eintr(
-    stream: &mut impl std::io::Read,
-    buf: &mut [u8],
-) -> std::io::Result<usize> {
-    loop {
-        match stream.read(buf) {
-            Err(e) if e.kind() == std::io::ErrorKind::Interrupted => continue,
-            other => return other,
-        }
-    }
 }
 
 fn read_pid_file(path: &str) -> u32 {
