+++ crates/vsock-guest/tests/connection/support/temp_paths.rs
@@ -1,0 +1,37 @@
+pub(crate) struct TempPathGuard(String);
+
+impl TempPathGuard {
+    pub(super) fn new(path: String) -> Self {
+        let _ = std::fs::remove_file(&path);
+        Self(path)
+    }
+
+    pub(crate) fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl Drop for TempPathGuard {
+    fn drop(&mut self) {
+        let _ = std::fs::remove_file(&self.0);
+    }
+}
+
+pub(crate) fn unique_tmp_path(label: &str, suffix: &str) -> TempPathGuard {
+    let nonce = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .unwrap()
+        .as_nanos();
+    TempPathGuard::new(format!(
+        "/tmp/vsock-test-{label}-{}-{nonce}{suffix}",
+        std::process::id(),
+    ))
+}
+
+pub(crate) fn unique_socket_path(label: &str) -> TempPathGuard {
+    unique_tmp_path(label, ".sock")
+}
+
+pub(crate) fn unique_pid_path(label: &str) -> TempPathGuard {
+    unique_tmp_path(label, ".pid")
+}
